server/requirements.txt
fastapi
uvicorn
python-dotenv
google-generative-ai
openai